Time: 0 ms (100%), Space: 13.4 MB (85.48%) - LeetHub

+++ 1464-maximum-product-of-two-elements-in-an-array/1464-maximum-product-of-two-elements-in-an-array.cpp
@@ -3,7 +3,7 @@ public:
     int maxProduct(vector<int>& nums) {
         sort(nums.begin(),nums.end());
         int n = nums.size();
-    long long max = (nums[n-1]-1)*(nums[n-2]-1);
+    int max = (nums[n-1]-1)*(nums[n-2]-1);
     return max;
     }
 };
